class08_20251005/06_chatbot_using_chainlit/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hassan Ali Khan", email = "73381969+hassan-ak@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.2",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[project.scripts]
chatbot = "chatbot:main"
agent = "agent.main:create_main_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
